Change how ClearInteract queue assigns chats to agents
Talks with ClearIt has resulted in us identifying a couple of issues.
To fix one if them the ClearInteract chat queue will now ask for
an agent resource for every CI task.
It is also now only possible for agents to get chat for the CI task
they responded to a request for.