Fix issue with how chat is assigned to agents
When the integration with Clear Interact is used then agents should
only be allowed to receive chat sessions in the areas which has
the same task identifier.

Previously when advanceQueue() was called then chat was assigned
to all agents who have access to the area.
Now only agents who have access to the area and do not have chat
in another area with another task identifier will be assigned
chat.