Retry errand queueing when Clear Interact returns error code -5
Error code -5 means that the Clear Interact task (task is what we call area)
is closed. In other words outside of the working hours.
Tasks are usually closed during the night and when this happens the system
should retry queuing errands until the tasks opens again.

List of error codes returned by Clear Interact for queueCentionItem.

 * -1 - specified "item_id" does not exist in the system (queued or assigned to agent).
 * -3 - is sent from the middle layer if there is a communication error
 * -4 - the task does not exist.
 * -5 - the task is closed.
 * -6 - the task is passive.
 * -7 - specified "item_id" is already queued.
 * -8 - specified "item_id" is assigned to agent.