Fix problem with errands bening queued multiple times in CI
If Clear Inteact respondes with error code -5 or -8 when when
invoking queueCentionItem or queuePersonalCentionItem then
the system will save a Workflow.ClearInteractAction to the database
that then will be picked up by a background script that will try to
invoke the function with the same parameters again until it succeeds.

An error can happen because the Clear Interact task (queue) is closed.

Lets say an agent is working directly in the Cention system outside
ofthe task's working hours.

Lets say a new errand comes in. It is outside of the task's working our.
An error is returned. An action object is created.
Lets say an agent is working directly in the Cention system.
Clicks on the errand in New errand and then decides to put back the errand in New errands.
Because it is outside the task's working hours an error will be returned and another action object will be created.

Once the task opens again then the same errand will be queued twice.

To stop this the code now checks if there already exists an action object before creating a new.