Fix handled errands being queued in Clear Interact
When the system tried to queue an errand but gets back certain
error codes, -5 is one of them which means that the task in CI
is currently closed then a Workflow.ClearInteractAction object
is created. This object contains all the paramaters that
the SOAP request for queing the errand received.

The idea is that from this point ClearInteractRunQueuedActions.fe
should use those parameters in that object to try to queue
the errand until successful.

So if the tasks closes at 8PM and doesn't open until 8AM the next
day then all those errands you came in after 8PM should get
queued because of the above behaviour.

The problem with handled errands being queued arises when agents
work in the system after 8PM even though the tasks are closed.
Previously the system didn't try to remove the Workflow.ClearInteractAction object
for an errand if it had such an object.
This mean that when tasks opened in the morning than already handled errands
would get queued in Clear Interact.

The previous fix for this wasn't 100% complete.
What the previous fix did was that it checked if a Workflow.ClearInteractAction object
existed when removeErrand() was called but the problem was that removeErrand()
was never called because queuedInExternal wasn't set to true when addErrand()
was called and created the Workflow.ClearInteractAction object.

The real fix now sets queuedInExternal to true whatever successful way
addErrand() can go. That means removeErrand() is called and that means
that Workflow.ClearInteractAction object is removed and that means
already handled errands shouldn't get queued in Clear Interact.