Stop errands from being queued twice or more i Solidus
The problem is that the status call doesn't make a difference
between no status because errand is not queued and no status because connection error.

That means if a temporary connection error occurs then the same errand will be queued
twice by SolidusStatusCheck.fe.

Add the possiblity to specify an error callback so that it is possible to know
when an error happens.